break up  provision and init scripts

+++ start-and-load.sh
@@ -5,26 +5,39 @@
 
 set -eo pipefail
 
-if [[ $# < 7 ]]; then
-    echo >&2 "USAGE: $0 --matrix-dir /path/to/matrices \
-                        --tree-path /path/to/tree \
-                        --label-path /path/to/labels \
-                        --port 1234 \
-                        [--debug]" && exit 1
+help()
+{
+   echo "Start the TMC containers and mount label and tree files, optionally loading in matrix data."
+   echo
+   echo "Syntax: start-and-load.sh --tree-path --label-path --port [--matrix-path] [--debug]"
+   echo "options:"
+   echo "tree-path   /path/to/cluter_tree.json."
+   echo "label-path  /path/to/labels.csv."
+   echo "matrix-path /path/to/matrices (files in matrix-market format, can be nested and/or gzipped)."
+   echo "port        <Port where webapp will listen>."
+   echo "debug       Print data import details."
+   echo
+}
+
+if [[ $# < 6 ]]; then
+    help && exit 1
 fi
 
 debug=""
 
 while [ ! -z "$1" ]; do
-    if [[ "$1" == '--matrix-dir' ]]; then 
+    if [[ "$1" == '--help' ]]; then
+        help && exit 1
+
+    elif [[ "$1" == '--matrix-dir' ]]; then
         shift
         matrix_dir=$1
 
-    elif [[ "$1" == '--tree-path' ]]; then 
+    elif [[ "$1" == '--tree-path' ]]; then
         shift
         tree_path=$1
 
-    elif [[ "$1" == '--label-path' ]]; then 
+    elif [[ "$1" == '--label-path' ]]; then
         shift
         label_path=$1
 
@@ -41,16 +54,16 @@ while [ ! -z "$1" ]; do
     fi
 done
 
-if [[ ! -d "${matrix_dir}" ]] ; then
+if [[ -n "${matrix_dir}" && ! -d "${matrix_dir}" ]] ; then
     echo >&2 "${matrix_dir} does not exist!" && exit 1
 fi
 
 if [[ ! -f "${tree_path}" ]] ; then
-    echo >&2 "${tree_path} does not exist!" && exit 1
+    echo >&2 "cluster_tree.json not found at ${tree_path}!" && exit 1
 fi
 
 if [[ ! -f "${label_path}" ]] ; then
-    echo >&2 "${label_path} does not exist!" && exit 1
+    echo >&2 "labels.csv not found at ${label_path}!" && exit 1
 fi
 
 if [[ -z $port ]]; then
@@ -59,9 +72,13 @@ fi
 
 docker-compose build
 
-docker-compose -f docker-compose.prod.yaml run --rm -p ${port}:3000 \
-    -v "${matrix_dir}":/usr/data/matrices:ro \
-    -v "${tree_path}":/user/data \
-    -v "${label_path}":/user/data \
-    node init $debug
+if [[ -n "${matrix_dir}" ]]; then
+    docker-compose -f docker-compose.prod.yaml run --rm -p ${port}:3000 \
+        -v "${matrix_dir}":/usr/data/matrices:ro \
+        node --init $debug
+fi
 
+docker-compose -f docker-compose.prod.yaml run --rm -p ${port}:3000 \
+    -v "${tree_path}":/usr/app/static/files/cluster_tree.json:ro \
+    -v "${label_path}":/usr/app/static/files/labels.csv:ro \
+    node --prod
